stm32l4-multi: makes: move libmulti external interface to separate dir
JIRA: CI-59
