+++ multi/stm32l4-multi/Makefile
@@ -6,13 +6,6 @@
 
 LOCAL_PATH := $(call my-dir)
 
-NAME := libstm32l4-multi
-SRCS := $(wildcard $(LOCAL_PATH)libmulti/*.c)
-HEADERS := $(wildcard $(LOCAL_PATH)libmulti/*.h)
-DEPS := libtty
-
-include $(static-lib.mk)
-
 NAME := stm32l4-multi
 SRCS := $(wildcard $(LOCAL_PATH)*.c)
 LOCAL_HEADERS := stm32l4-multi.h
