+++ multi/stm32l4-multi/libmulti/Makefile
@@ -1,0 +1,13 @@
+#
+# Makefile for Phoenix-RTOS stm32-multi
+#
+# Copyright 2022 Phoenix Systems
+#
+
+LOCAL_PATH := $(call my-dir)
+
+NAME := libstm32l4-multi
+SRCS := $(wildcard $(LOCAL_PATH)*.c)
+DEPS := libtty
+
+include $(static-lib.mk)
