+++ multi/stm32l4-multi/libmulti/include/libmulti/libi2c.h
