+++ multi/stm32l4-multi/libmulti/include/libmulti/libspi.h
